net: add IPv4 routing FIB support for swdev
To offload IPv4 L3 routing function to swdev device, the swdev device driver
implements two new .ndo ops (ndo_sw_parent_fib_ipv4_add/del).  The ops
are called by the core IPv4 FIB code when installing/removing FIB entries
from the kernel FIB.  On install, the driver should return 0 is FIB
entry (route)
can be installed to device for offloading; -EOPNOTSUPP if route cannot be
installed due to device limitations; and other negative error code on failure
to install route to device.  On failure error code, the route is not installed
to device, and not installed in kernel FIB, and the return code is propagated
back to the user-space caller (via netlink).

The FIB entry (route) pushed down to device/driver is scanned for nexthops.
The nexthops referenced by the device must all be ports on the same swdev
device.  (If not, the swdev can't offload this route because it's spanning ports
in and out of the swdev).

Since the FIB entry is "naked" when push from the kernel, the driver/device
is responsible for resolving the route's nexthops to neighbor MAC addresses.
This can be done by the driver by monitoring NETEVENT_NEIGH_UPDATE
netevent notifier to watch for ARP activity.  Once a nexthop is resolved to
neighbor MAC address, it can be installed to the device and the device will
do the L3 routing in HW.

Signed-off-by: Scott Feldman <sfeldma@gmail.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>